requirements.txt
# Multi-Property Hotel PMS MVP
Flask>=3.0
Flask-SQLAlchemy>=3.1
Flask-Migrate>=4.0
Flask-Login>=0.6
Flask-WTF>=1.2
Flask-Limiter>=3.5
psycopg2-binary>=2.9
python-dotenv>=1.0
Werkzeug>=3.0
